requirements.txt
# Core dependencies
pydantic>=2.0.0
pyserial>=3.5
python-dotenv>=0.19.0
requests>=2.28.0

# QR Scanner HID support (Linux only - required)
evdev>=1.4.0; sys_platform == "linux"
